Add ErrorEvent class for clearer modeling
