+++ sentry-ruby/lib/sentry-ruby.rb
@@ -12,6 +12,7 @@ require "sentry/utils/logging_helper"
 require "sentry/configuration"
 require "sentry/logger"
 require "sentry/event"
+require "sentry/error_event"
 require "sentry/transaction_event"
 require "sentry/span"
 require "sentry/transaction"
