+++ sentry-ruby/lib/sentry/transaction_event.rb
@@ -1,32 +1,15 @@
 # frozen_string_literal: true
 
 module Sentry
+  # TransactionEvent represents events that carry transaction data (type: "transaction").
   class TransactionEvent < Event
     TYPE = "transaction"
-
-    SERIALIZEABLE_ATTRIBUTES = %i(
-      event_id level timestamp start_timestamp
-      release environment server_name modules
-      user tags contexts extra
-      transaction platform sdk type
-    )
-
-    WRITER_ATTRIBUTES = SERIALIZEABLE_ATTRIBUTES - %i(type timestamp start_timestamp level)
-
-    attr_writer(*WRITER_ATTRIBUTES)
-    attr_reader(*SERIALIZEABLE_ATTRIBUTES)
 
     # @return [<Array[Span]>]
     attr_accessor :spans
 
-    # @param configuration [Configuration]
-    # @param integration_meta [Hash, nil]
-    # @param message [String, nil]
-    def initialize(configuration:, integration_meta: nil, message: nil)
-      super
-      @type = TYPE
-      self.level = nil
-    end
+    # @return [Float, nil]
+    attr_reader :start_timestamp
 
     # Sets the event's start_timestamp.
     # @param time [Time, Float]
@@ -39,6 +22,7 @@ module Sentry
     def to_hash
       data = super
       data[:spans] = @spans.map(&:to_hash) if @spans
+      data[:start_timestamp] = @start_timestamp
       data
     end
   end
